Trekk ut gjenbrukbar visLagreKnapp funksjon
